refactor mnist
